move attribution_class type into artwork
